requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
selenium>=4.1.0
webdriver-manager>=3.8.0
flask>=2.2.2
flask-wtf>=1.0.1
openai>=0.27.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pyinstaller>=6.5.0
PyGithub==2.1.1
